add homepage skeleton-basic design
