+++ Frontend/package.json
@@ -21,6 +21,7 @@
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
+    "sass": "^1.70.0",
     "vite": "^5.0.8"
   }
 }
