+++ Frontend/src/pages/homepage/Homepage.scss
@@ -1,0 +1,24 @@
+
+.homepage{
+    display: flex;
+    gap:80px;
+    margin-top: 100px;
+.homepage-titles{
+    line-height: 1.5;
+    
+}
+.menu {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 20px;
+    
+    .menu-button {
+        height: 200px;
+        width: 200px;
+        border-radius: 20px;
+        background-color: rgb(0, 0, 71);
+        color: white;
+        font-size: 24px;
+    }
+}
+}
